Define symbols.zip build rules in Soong
symbols.zip file consists of all copies of LOCAL_SOONG_UNSTRIPPED_BINARY
entires in $PRODUCT_OUT/symbols of modules that are installed. This
change utilizes allInstalledModules of android_device module to collect
the unstripped binary files of all installed modules.

This change generates the build rule of the symbols.zip and the merged
elf mapping textproto file, and dists them.

Test: m droid dist &&
 diff out/dist/<product name>-symbols-<user name>.zip with/without this change
Bug: 395989947
Change-Id: I5552bc91b1c6a0c4c29a98705867b430125a41b2