soong: Run installclean for gapps earlier than usual
Globbing behavior is different now
Run installclean for gapps or variant switch before it happens to restore old behavior
Also move config file from gapps to common outdir instead of product out. This is done so we installclean regardless of the device we are building, when we switch gapps flag.

test: builds and verified build output
Change-Id: Ib851721fd4330ee83783de4ee9cbb20c8df094c9