A mostly functional kotlin incremental client.
Still a little rough, but this code can successfully compile
Android when integrated into Soong (see follow up cl).

This code does not yet track file changes correctly between
incremental runs, so it is _still_ not suitable for actual use,
but that is the only significant missing piece.

After that, logging needs to be improved, but we're otherwise
in a good place!

Bug: 392906864
Test: SOONG_USE_PARTIAL_COMPILE=true SOONG_PARTIAL_COMPILE=true,+Enable_incremental_kotlin SystemUI (along with integration into soong)
Change-Id: I3ce91c8cb08fa9a6f3759ad4e21bb3b1b3191f23