Enable sparse encoding for AAPT2 in Soong.
Leads to smaller APKs, especially after "resource shrinking" (which
leaves holes in the resource table).

There are multiple paths that apply this optimization: the aapt2 link,
convert, and optimize commands. In Soong, there are only rules for link
and convert. Apps will normally be linked via the link rule, but in when
resource shrinking is enabled, they will additionally go through
convert. In the former case, it would be sufficient to add
--enable-sparse-encoding to just the link stage, but in the latter case
it also needs to be specified in convert, or else the sparse encoding
will be undone.

I've updated the aapt2convert rule to include --enable-sparse-encoding,
and I added --enable-sparse-encoding to the linkFlags that are passed to
the aapt2link rule. I tested that ClockworkSetupWizard.apk comes out the
end with sparse encoding enabled on seluna-trunk_staging-userdebug.

Bug: 374251712
Bug: 123358928
Test: Build with and without flag, verify sparse encoding in APK
Flag: build.RELEASE_SOONG_SPARSE_ENCODING

Change-Id: Idab84153d537e81e7284ff0415645a1b8d12d6e6