Update test_packages for converting continuous_instrumentation_tests
Introdue a new attribute for test_packages to add tests for the first
supported arch variant if exist.
Also update the test_packages dependencies package rule, always packed
apps and frameworks if they are direct dependencies of the
test_packages.

Ignore-AOSP-First: Because the continuous_instrumentation_tests
conversion is in internal branch currently.

Bug: 399246722
Test: m dist continuous_instrumentation_tests
Change-Id: I7b18d22ed26c5a9831846f3cbfa416779cc0fec5