Put AArch64 branch protection flags in a feature.
I need to add ARMv8.5-A and ARMv8.7-A and the duplication
was getting out of hand. Instead of specifying these as arch
variant flags, specify them as arch feature flags for a new
feature called "branchprot" that is enabled on the relevant
variants.

This also fixes an inconsitency where armv8-a-branchprot
enabled branch protection, but didn't disable the stack
protector, which appears to have been an omission.

Bug: 364383722
Test: inspect compilation flags in a local build, presubmit
Flag: EXEMPT infeasible
Change-Id: Ib5f89b5f0f7a597118edaf2b7a70dbad515616f5