Reapply "Always install test files with soong"
This reverts commit a933f952666f1282452da43ed5b73ebbdc091b37.

Previously, this commit included a change where I prioritized the fact
that it was a test over the fact that the app was priviledged for
installation:
https://googleplex-android-review.git.corp.google.com/c/platform/build/soong/+/32747769/2/java/app.go#999

I did this because I was worried about modules already installing
files that this new testcase installation code would then try to
install. I added a check for already installed files in test_suites.go.
I thought that I would have to tell test_suites.go about the already
installed file via the TestSuiteInfo, but I actually ended up getting
that information from the CommonInfo. Using the installed file as the
source file is problematic in soong+make builds, because make adds a
dependency from the main installed file to the testcase files, and if
the testcase files also have a dependency to the main installed file,
it's a circular dependency. Swapping the priv-app / testcases priority
worked around the circular dependency, but it turns out it causes later
issues where some tests are not installed with the privileges they need
to run correctly.

Bug: 388850000
Test: Presubmits
Change-Id: I686c1a7ba4f2367de7a75bf7052cb54d4296265c