Don't glob genrule dir_srcs
The trusty dir_srcs do a lot of very large globs, leading to extra
Soong reanalysis when any files are added to large parts of the tree.
Soong doesn't need to know the list of files during analysis, it only
needs the list of files as dependencies to rerun the genrule.  This
can be handled with a depsfile instead.

Add a new dir_to_depfile tool that produces a list of all of the
subdirectories and files in a directory in depfile format, and
add it to the command line when using nsjail + dir_srcs.

Bug: 375551969
Test: builds
Change-Id: I9b2830919b5325f700cc27d9c602b45f9fed3707