Filter flag pb file before generating android_flags.jar
This change makes the rule of exported_java_aconfig_library filtering
the flag pb files first to remove flags that are exported but not used
for any API, and then pass the filtered pb files to aconfig to generate
android_flags.jar.

After this change, only exported flags used for API will be exposed to
android_flags.jar.

NO_IFTTT=change the location of the condition line

Test: m exported_java_aconfig_library
Bug: 345798636

Change-Id: Id0a342ae1643c12d5954f4b86a4604bd653f4c92