Introduce Soong api DistForGoalWithFileNameTag(...)
The newly introduced api works similar to DistForGoalWithFilename, but
appends the file name tag placeholder suffix, which is later converted
to the unique host identifer (build id for ci builds, username for local
builds).

This change also converts the existing dists with the file name tag
placeholder suffix to using the new api.

Test: m droid dist --soong-only
Bug: 407562438
Change-Id: I3e1b80bd1bfe4f89beb6c2ef2004580a909df13d