Build device-platinum-tests.zip using Soong
The packaging rules are similar to performance-tests.zip, which was
added to Soong in ag/32749281.

`buildPerformanceTests` has been renamed to `packageTestSuite`, and an
enum has been added to distinguish performance-tests and
device-platinum-tests. device-platinum-tests also dists a zip containing
host .so files, and a new rule has been added to build that.

Bug: 405183828
Test: Soong built device-platinum-tests.zip is bit-identical to Make
built device-platinum-tests.zip

Change-Id: I2c15b064f77d2ed1876f6e84820d8d6a360c635b