rust: Ensure rustdoc implicits are available
The variant that builds rustdocs may vary based on which mutators are
applied to a module. Prior to stub support being added, all variants
called `rustc` initially so dependencies of rustdoc were being built in
that step.

With stub support added, stub variants do not call `rustc`. If the
rustdoc producing variant is a stub variant, this leads to rustdoc
failing because implicit deps haven't been built.

We can resolve this by explicitly listing rustdoc dependencies as
implicit in its build rule.

Bug: 403635935
Test: m rustdoc
Change-Id: I950644b6743a9f45c5bd53e89fdbf24c8dbecad3