Track modules for tests in Soong
A future change to Blueprint will allow it to drop its last references
to Modules after their GenerateBuildActions is complete.  That will
prevent tests from retreiving the Module objects with ModuleForTests.
When running with a test Config, keep a list of all of the modules
that ModuleForTests can iterate over.

Test: all soong tests pass
Change-Id: I2dc7696887c49d8b25bb55a0116c4abdc5e7350b