Wrap blueprint.Context.Visit* for android.TestContext
Add wrappers in android.TestContext that convert the
blueprint.Context.Visit* methods to using android.Module and
android.ModuleProxy.  This reduces the mixing of blueprint.Module
and android.Module in Soong code, adn will also be used to add
extra logic to VisitAllModules and VisitDirectDeps later to convert
ModuleProxies to Modules in tests once Blueprint only has a
ModuleProxy available.

Test: all soong tests pass
Change-Id: Id497dc7caf4ac5fd2c2c23d27c17e124bec6ab8d