Skip phony target generation for Androidmk hidden modules
Prior to ag/32854612, the renamed module name was used as the phony
name. In the case of apache-xml, we had two phony targets
```
phony apache-xml # build source
phony prebuilt_apache-xml # build prebuilts
```

The latter is intentionally unbuildable. It checks in a corrupt file so
that rdeps that do not inadvertently compile against the implementation
library from prebuilts/module_sdk, which is not available (module sdk
contains stubs).

After ag/32854612, the phony changed to
```
phony apache-xml # build source and prebuilts
```

This CL still uses BaseModuleName (name as it appears in Android.bp),
but does so only if the module is exported to make. This makes it
closer to the make implementation

Bug: 409225101
Test: lunch aosp_cf_x86_64_phone-trunk_staging-userdebug
Test: aninja -t path apache-xml out/soong/.intermediates/prebuilts/module_sdk/art/current/sdk/prebuilt_apache-xml/android_common_prebuilt_com.android.art/local-combined/apache-xml.jar
(path no longer exists)

Change-Id: I267b07505cc4a0b70f70c635f34f646dc3b40afd