Add SOONG_PARTIAL_COMPILE to banner
This adds SOONG_PARTIAL_COMPILE to the banner when the user has
requested that it be used with `SOONG_USE_PARTIAL_COMPILE=true`.

Additionally, disable partial compile for some targets (including
`droid` and `sync`) since some of the partial compile speed
optimizations cause the image to become too large for many devices.
If we are not dumping variables (such as during `lunch`) and we have
temporarily disabled partial compile, say that in the banner.

Bug: b/374975543
Bug: b/409810224
Test: manual, TH
Change-Id: Id1a485df1b01a624d6e870d039e5a1b1f0017228