Use coverage deps in test_package
When coverage is enabled, non-coverage variants don't install/package
any files, and the coverage variants do. Our test package was always
depending on the non-coverage variants, so when coverage was enabled
they weren't installing anything and the test package was mostly empty.

The coverage mutator supports modules implementing
IsNativeCoverageNeeded() to indicate that they want the coverage
variants of their dep, so use that.

Bug: 408492405
Test: CLANG_COVERAGE=true m continuous_instrumentation_tests_soong, check that it has the same files in it as the non-coverage variant
Change-Id: I01d068d9630518c4bcf62c0362fbfa42e0f71311