Add impl and header jars to checkbuild of DeviceForHost/HostForDevice
The impl and header jars of java_device_for_host and
java_host_for_device are in the transitive closure of the phony
$(module_name) target in Make builds. soong_java_prebuilt.mk adds them
to deps of LOCAL_BUILT_MODULE which is itself added to deps of the phony
(host|target)_$(module_name)_all_targets.

These jars are currently not in the transitive closure of $(module_name)
in soong only builds due to which building such modules in `m
--soong-only` mode does not build anything. This CL fixes it by adding
these jars to checkbuild.

Bug: 410693901
Test: lunch aosp_cf_x86_64_phone-trunk_staging-userdebug (soong-only
turned on by default)
Test: m nothing
Test: aninja -t path robolectric_android-all-device-deps out/soong/.intermediates/external/robolectric/robolectric_android-all-device-deps/linux_glibc_common/combined/robolectric_android-all-device-deps.jar
(verified path exists)

Change-Id: If4152070eb3399d34734abd5ac4d888ee9a3beff