Don't copy images to target-files.zip
add_img_to_target_files will build the images again from the files
that are already in target-files.zip. When it does so, it sets a fixed
timestamp for the file mtimes so that they're hermetic. We were also
doing that in the filesystem modules, but we weren't doing it for
the vbmeta modules, causing a mismatch and a boot failure.

Bug: 376727180
Test: ./verify_img_zip.sh out/soong/.intermediates/build/soong/fsgen/aosp_cf_x86_64_only_phone_generated_device/android_x86_64_silvermont/img.zip where verify_img_zip.sh is:

```

tempdir=$(mktemp -d)

unzip $1 -d ${tempdir}
(
    cd ${tempdir}
    simg2img super.img super_unsparse.img
    lpunpack super_unsparse.img
    for i in *_a.img; do
    cp ${i} ${i/_a.img/.img}
    done
    avbtool verify_image --image vbmeta.img --follow_chain_partitions
)
```

Change-Id: I80c45f16fddc4f737bb95164214014235faf3da7