Handle fsgen partition of override apps
If the base app is not listed in PRODUCT_PACKAGES.

The previous implementation in ag/33086725 assumed that if an override
app is in PRODUCT_PACKAGES, the base app is in PRODUCT_PACKAGES as well.
This is not always true. To support this case, a new prop has been added
to FsGenState that can be used to lookup partition of the base module.

Test: go test ./fsgen
Bug: 410570903

Change-Id: I9314b3e52db9e54be6bcae0ab556a1fa5a5a2ecf