Always use file_contexts_bin_gen in target-files
Previously we were using the system image's selinux context. When
we autogenerate partitions, we always use file_contexts_bin_gen as
the selinux context. But with the handwritten soong image, they use
a different selinux context that only has the selinux information for
the system image. file_contexts_bin_gen has information for all images.

Just hardcode the usage of file_contexts_bin_gen for the
target-files.zip package. This is the same as what make does.

This selinux contexts file is later used to build the "hermetic"
versions of the images in the target-files.zip package, so it's
important it has info for all partitions.

Bug: 376727180
Test: ./prebuilts/build-tools/linux-x86/bin/py3-cmd build/soong/scripts/soong_only_diff_test.py aosp_cf_x86_64_phone
Change-Id: I75a82a035fd76ccd5c0621316a1735e504d50c29