Don't add a namespace prefix if the namespace is exported to make
Soong-only builds would always add the soong namespace as a prefix to
the module target if the module was in a namespace. This doesn't match
what make does. Make would never add a namespace prefix, but it would
also only see / generate a phony target for the modules in namespaces
that are listed in PRODUCT_SOONG_NAMESPACES. Replicate this
functionality in soong.

I've also kept the namespace-prefix targets regardless of if the
namespace is in PRODUCT_SOONG_NAMESPACES or not.

Bug: 411694367
Test: m libg3a.af_test
Test: diffed the ninja files, the only differences are in the -packaging ninja file, as more phonies are created now.
Change-Id: Ie9202ef753ffe34a2ed72dc89d08ca68a92750df