Ignore bootclasspath libraries in manifest_check.
A shared library may be added to the bootclasspath since a certain
platform version. Apps' manifests may still include such a library, to
maintain compilatibility with old platforms, especially when the app is
built elsewhere and dropped as a prebuilt, not being part of the
platform build.

Soong filters out bootclasspath libraries when constructing CLC, and
passes the filtered CLC to manifest_check, causing an expected
discrepancy between the build time value and the manifest value. This
change fixes it by filtering out bootclasspath libraries from the
manifest value, to match the build time value.

Bug: 413330617
Test: m
Flag: EXEMPT bugfix
Change-Id: Ic46909c096e208e3fc4cee99cf298f571e9bc269