Track deps on build fingerprint/thumbprint files
In make builds, we rely on `m installclean` to delete a bunch of files
and force them to be regenerated. CI always runs `m installclean`, but
locally it's often not run. In soon-only builds, we
try to avoid requiring that step, so we more precisely track
dependencies. However, we don't want to cause excessive rebuilds still,
so we don't add dependencies on stuff like the build date.

Previously, I think I was concerned that the build fingerprint/
thumbprint files would change every build and cause excessive rebuilds,
so I didn't add them as dependencies. And they actually did use to
change every build, and still do on ci, because they contain the build
number. But we've changed it so that now local builds don't include
the build number, so it doesn't cause rebuilds locally, but still will
on CI. And we do want the rebuild behavior on CI in order to update
the build number.

There was a bug reported where the build number was not being updated
correctly, due to the lack of rebuilds.

Bug: 411231354
Test: m out/soong/.intermediates/build/soong/fsgen/aosp_cf_x86_64_only_phone_generated_device/android_x86_64_silvermont/target_files.zip, check the build number in META/misc_info.txt and ODM/etc/build.prop. Then change the build number and rebuild, ensure both of those files get the new build number. Then rebuild without changing the build number, ensure you get the "no work to do" message.
Change-Id: I0d5ada76da4af6465c71c3e0b000a480c647f2fe