Do not emit non-dependency required entries to Androidmk
Required deps are not identical across variants. Some dep may be added
as a required dep in a device variant, but may not be added as a dep in
a host variant. However, when generating the Androidmk entries Soong
currently does not distinguish (*_)required entries that are actually
added as deps to the ones that are not added as deps for each variants,
but naively pass the entries listed in the module definition to make.

This may lead to discrepancies in required deps between Soong only build
and Soong plus make build, given that the required deps are added as
order only dependencies in make. Such discrepancies lead to rebuild of
the target (even for soong modules) between soong only vs no soong only
toggle as the dependencies change.

This change fixes the issue by only listing the modules that are
actually added as deps when generating the `LOCAL_*_REQUIRED_MODULES`
Androidmk entries.

Test: build/soong/scripts/soong_only_diff_test.py aosp_cf_x86_64_phone
Test: diff out/soong/Android-aosp_cf_x86_64_phone.mk before and after
Bug: 413097935
Change-Id: I7fd55f22cb8067545c2970e65c93a25dfd762a32