Move the installation rules of fuzz shared deps to Soong
Shared lib deps of fuzz binaries (sanitized or non-sanitized) are
installed by Make to the following locations
- (HOST_OUT|PRODUCT_OUT)/data/fuzz
- PRODUCT_OUT/symbols/data/fuzz (device)

This is currently done by a singleton that collects all the dependencies, dedupes
and emits this information to Make for copy-many-files. The CL updates
the implementation to the following
- Use InstallPairs instead of RuleBuilderInstalls to collect the
  information in the singleton. The singleton will then create a
CpIfChanged rule for each pair.
- Add the dst in the pairs to the installdeps of the fuzz'd binary. This
  ensures that building the binary will install the fuzz deps. We can
now remove `LOCAL_FUZZ_INSTALLED_SHARED_DEPS`.

Test: m nothing --no-skip-soong-tests
Test: m haiku
Bug: 414484738

Change-Id: If4ae13dee86bbd5cea3339568fbc559ee5e23cf3