Propagate transitive CLC of SDK libraries from "libs".
Similar to Ief74725d48853b400f2c7dc3b03eb6a349feb7c9, but for "libs".

If an app depends on SDK library "foo", and "foo" depends on "bar"
through `libs: ["bar.impl"]"`, "bar" should be added to the CLC of the
app.

This functionality had been there for java_library, but not for
java_sdk_library. This fix adds the functionality for java_sdk_library,
to eliminate the inconsistency.

Bug: 414306949
Test: m --no-skip-soong-tests nothing
Flag: EXEMPT bugfix
Change-Id: I7280c1ea5d35089fdcc4c9682f07cfed898fb81d