Add "uses_libs" to the CLC of SDK libraries.
If an SDK library "foo" depends on "bar" through `uses_libs: ["bar"]"`,
"bar" should be added to the CLC of foo.

This functionality had been there for java_library, but not for
java_sdk_library. This fix adds the functionality for java_sdk_library,
to eliminate the inconsistency.

Bug: 414306949
Test: m --no-skip-soong-tests nothing
Flag: EXEMPT bugfix
Change-Id: I8fb780db1c073a0b3f92685bd544b9a32d3874f8