rust: Fix Generated Rust Staticlibs on Windows
This fixes generated staticlibs for windows targets, addressing two
bugs. The first is that lib and pdb files should only be generated
when the build target is a shared library, not the module context.

The second is that the `-l` flags that are needed for Rust windows
targets also need to be added to cc modules as they may include Rust
generated staticlibs.

Bug: 414568551
Test: m <cc module with Rust staticlibs>
Change-Id: I52e8774be3f282ee5ea05a558bbd5d5b5718f79b