Lower build niceness from 19 to 5
nsjail defaults to setting the nice value to 19, the lowest possible
priority.  This is lower than all the background monitoring-type
processes, which tend to be set to 10.  Lower the niceness from 19
to 5, which is still lower priority than the default value of 0.  This
should allow the UI to remain responsive while a build is running
while allocating more CPU to the build than to background processes.

Test: builds
Change-Id: I2718aea2ecbc47f454797d7d7e5c35963d270895