Handle variants for unbundled_builder in transitions
unbundled_builder was trying to guess the variants of the apps in
TARGET_BUILD_APPS. This kindof works, but was errorprone. Instead,
make each individual transition mutator handle the decision for what
variant the unbundled_builder should use. This allows deciding based
on what variants are available.

The motivating example for this change was libextservices_test-tplus,
which sets `compile_multilib: "prefer32"`. The unbundled_builder itself
doesn't know that, so it was getting a missing dependency error as
the unbundled_builder module used a 64 bit arch.

Bug: 413145177
Test: TARGET_RELEASE=mainline_2025_07 TARGET_BUILD_VARIANT=userdebug PRODUCT=mainline_modules_x86_64 COVERAGE_MODULES=EXTSERVICES_TPLUS ./vendor/google/build/build_unbundled_coverage_mainline_module.sh -j64
Change-Id: I574150d7006a384b5569b78e98068e6407f05fbe