Build the unbundled jacoco report zip with soong
For soong-only unbundled builds.

This does produce some diffs in the naming of files in the zip.
For modules that are contained within an apex, previously they'd be
under paths like "MyModule.com.android.myapex", but now they're just
under "MyModule". This is because each file in the apex gets a make
module that has the apex name as a suffix (generated in
androidMkForFiles). We don't really want these modules to exist, so I'm
removing this remneant of them. However if it becomes problematic it
should be straightforward to maintain them.

```
1a2,3
> out/target/common/obj/APPS/GoogleExtServices-sminus_intermediates/jacoco-report-classes.jar
> out/target/common/obj/APPS/GoogleExtServices_intermediates/jacoco-report-classes.jar
5,7c7
< out/target/common/obj/ETC/GoogleExtServices-sminus.com.google.android.extservices_intermediates/jacoco-report-classes.jar
< out/target/common/obj/ETC/GoogleExtServices.com.google.android.extservices_tplus_intermediates/jacoco-report-classes.jar
< out/target/common/obj/JAVA_LIBRARIES/android.ext.adservices.com.google.android.extservices_intermediates/jacoco-report-classes.jar

Change-Id: I9eb53c8a65f18e220818cc58c0a0b54d707f5ef3
---
> out/target/common/obj/JAVA_LIBRARIES/android.ext.adservices_intermediates/jacoco-report-classes.jar
```

Bug: 413145177
Test: TARGET_RELEASE=mainline_2025_07 TARGET_BUILD_VARIANT=userdebug PRODUCT=mainline_modules_x86_64 COVERAGE_MODULES=EXTSERVICES_TPLUS ./vendor/google/build/build_unbundled_coverage_mainline_module.sh -j64, diffed disted jacoco zips
Change-Id: Id6349a60fffbdd3e49f9ac2ec2fd91e6f2644ca4