Install autogenerated rros of vendor and product apps
The autogenerated soong vendor and product partitions visits the
transitive closure of system and system_ext to find autogenerated rros,
but does not currently look at possible rros of its own deps. This
means that if an app is product_specific and has autognerated rros, it
was previously missing from soong autogenerated filesystem.

Test: verified locally that the autogenerated rro in b/417231812#comment2 is no longer missing
Bug: 417231812
Change-Id: I66f9ac6bd602d3f3b7389015244843a65e37b16a