Flag controled clang version
Use flagging to control the version of Clang being used to build the
Android platform. This allows us to push new Clang versions gradually to
different build tracks.

The original ClangDefaultVersion, ClangDefaultShortVersion variables are
kept for now due to many users (including vendor and third party code)
that depend on the flag. They will be removed in a future version.

Test: presubmit
Bug: 364956472
Change-Id: Ic557a002b15c30a1ce1a7e2fc05445f9df3d7fde