Do not add vndk apexes as deps of autogen filesystem module in riscv64 builds
Currently, com.android.vndk.v* modules are automatically added as deps
of the autogenerated system_ext filesystem module, because the macro
`auto-included-modules` in make adds the same modules to the list of
installed modules.

However, these modules are disabled in riscv64 arch builds, and are not
exported to make. In other words, these modules do not exist in make in
riscv64 builds and they are simply skipped. To align with such behavior
in soong only build, this change disables the automatic dep
addition of com.android.vndk.v* modules in riscv64 builds.

Test: lunch aosp_cf_riscv64_phone-trunk_staging-userdebug && m
Bug: 417233553
Change-Id: I729e9a194968ce1595a337aa08105ab003a039b0