Build apkcerts.txt with soong
Make will still handle building apkcerts.txt for make modules, but
soong will build it for soong modules and then make will combine
the make and soong apkcerts files.

This is for soong-only builds. In a followup change I'll make it so
that only soong dists the apkcerts.txt file on unbundled builds.

Bug: 413145177
Test: Presubmits, make will doublecheck that the soong-built apkcerts.txt is correct
Change-Id: I7aa011b27056adb4a16106ae1e040b7b3f184b51