Build abidiff as a validation action of the library
Previously the abidiff was only built at part of
LOCAL_ADDITIONAL_DEPENDENCIES, which won't work on soong-only builds.
Also as of ag/33582753 it's not even built on soong+make builds
currently because we made them more similar to soong-only builds.

Insead, add a validation dependency from the built library to the
abidiff results.

Bug: 419663728
Test: aninja -t path out/target/product/<product>/vendor/lib64/libutils.so out/soong/.intermediates/system/core/libcutils/libcutils/android_vendor_arm64_armv9-2a_cortex-a55_shared/libcutils.so.opt0.abidiff
Change-Id: I8639e12247a5bcdedbed7ad439fb995cadc2c9a3