Reapply "Build apkcerts.txt with soong"
This reverts commit ea7a009d871516d157711ff84d4ae338fb1c0985.

Make will still handle building apkcerts.txt for make modules, but
soong will build it for soong modules and then make will combine
the make and soong apkcerts files.

This is for soong-only builds. In a followup change I'll make it so
that only soong dists the apkcerts.txt file on unbundled builds.

Bug: 413145177
Test: Presubmits, make will doublecheck that the soong-built apkcerts.txt is correct
Change-Id: I1d958cec16ccb517df456c5d27312a63fac93316