Use target product arch instead of host arch in defaultDepCandidateProps
defaultDepCandidateProps(...) is used for specifying the modules that
are added as deps of the filesystem module regardless of the target
product by default. Currently, the arch of the generated default prop is
config.BuildArch, which represents the arch of the host, not that of the
target product. This becomes an issue when the bitness of the host and
the target product differs (e.g. when soong-only building a 32 bit
target product on a 64 bit host).

To resolve the problem, this change modifies the arch of the default
prop to that of the target product, so that the modules added as deps by
default are not erroneously elided due to bitness mismatch but always
installed as expected.

The issue has not been surfaced until now as all soong only enabled
products so far were 64 bit targets, and the host on CI are all 64 bit
targets. However this becomes an issue when adding the soong only build
support for 32 bit arch products.

Test: build/soong/scripts/soong_only_diff_test.py cf_gwear_x86
Bug: 421944148
Change-Id: Ic848f8db9b0547ad5725871f23c48c301fe0d286