Create a partial implementation of check-vintf-all for soong-only
Make generates checkvintf rules per partition (system, vendor) and a
monolithic checkvintf rule per device. This CL implements the per
partition checkvintf rule.

It will use checkvintf's dirmap to map $partition to the staging dir of
the soong filesystem. Similar to make, apexd_host will be used to
extract apexes to an intermediate directory to ensure that the apex
vintfs are not elided.

This implementation has a limitation: checkvintf needs to wait for the
staging directory to be assembled in the soong implementation. In make,
checkvintf needs to wait for the apexes and the vintf manifests
installed in the partition.

Bug: 415130821
Test: lunch aosp_cf_x86_64_phone-trunk_staging-userdebug
Test: m check-vintf-all
Test: check_vintf_system.log and check_vintf_vendor.log are not
bit-identical since the log contains the vintf paths, but the word count
is identical

Change-Id: Ic5b794ead990bb0d251c1573cd88927770a676ac