Configure RBE rust pool based on build variant
In `eng`, we pass codegen-units=16 to rustc, which takes advantage of
more parallelization, while in other builds we use codegen-unit=1. We
currently maintain two RBE pools, `java16` with e2-standard-16 machines,
and `default` with e2-standard-2 machines.

On userdebug builds, java16 machines provide somewhere around a ~25%
worker time savings, which isn't all that significant for a machine
taking 8x the resources. (This does not equate to a 25% savings in total
build time, just the individual uncached actions)

On eng builds, java16 machines provide a ~50% savings, and default
machines are actually a bit slower than userdebug builds. While this
isn't 8x, it's still a reasonable use of the larger pool.

Bug: 407072297
Test: ninja file has the appropriate RERustPool set for both eng and
  userdebug
Change-Id: I7e0e4703be39a0283fbfd6ac693d1044ed4cffbe