Do not install auto gen rros of apex in soong only build
In make, auto generated rros are only installed for the platform variant
of a module. In other words, auto generated rros of an apex-embedded
apps are not installed. Currently in soong only build such distinction
between the platform app and the apex app does not exist.

This change adds such distinction by stop traversing deps in
addAutogeneratedRroDeps mutator if source/prebuilt apex bundle module is
encountered while traversing deps.

Note that this change also moves the addAutogeneratedRroDeps mutator
from a post deps mutator to a post apex mutator, so that it can utilize
the apex providers set in apex package post deps mutators.

Test: build/soong/scripts/soong_only_diff_test.py cf_x86_64_foldable
Bug: 421945977
Change-Id: Idad185ef487bd0dc47e4ad056c1950c23e762a4f