Soong only diff test - append entry even if extra info is missing
Soong_zip occasionally skips appending extra data to symlink entries,
even if the symlink is not broken. Thus, skipping entries that do not
contain extra data fields may lead to omitting valid symlinks in
comparing the zip entries map between the soong only build and the soong
plus make build.

This change loosens this requirement and allow symlinks with empty extra
data can also be added to the zip entries map. Note that symlinks will
continue to be added to the map with its target as the value, thus
symlinks with different targets between soong only build and soong plus
make build will continue to be detected by the script.

Test: build/soong/scripts/soong_only_diff_test.py cf_x86_64_auto
Bug: 421945305
Change-Id: I6504d3282b62685d4ac414b46c62f681f7204cc5