build-flag: fix unexpected return error
In the `GetDefaultMapPaths` function, the return variable `err` is
re-used other places in the function. However, there is a logic checking
whether to add certain path based on if that file exists on the file
system, and the return error variable is re-used here for checking; if
the err is not nil, that just shows we should not add that path rather
than there is an error returned by this function.

Declare a new error variable in that checking instead of re-using the
return variable.

Bug: None
Test: manual
Change-Id: I11d4f20c140f4087f852a664eb52af96ddccbc63