Allow droidstubs compatibility check to be disabled
Previously, specifying `check_api.last_released.api_file` would always
cause a compatibility check to be performed. That was a problem because
it was also used to supply the previously released API to which flagged
APIs could be reverted. In that case there was no need to perform a
compatibility check.

This change adds an `enabled` property that, when set to `false` will
prevent the compatibility check from being performed.

Bug: 425309457
Test: ./gradlew
Change-Id: I99d5f4e7c3bbebdbf49d1857fbbbdb2d2fdd1fdc