Use empty previous API when unsafe_ignore_missing_latest_api: true
Previously, when `unsafe_ignore_missing_latest_api == true` then Soong
would not pass any previous released API to Metalava via the
`--check-compatibility:*` command line options. That meant that
Metalava could not fail if it processed flagged APIs without a
previously released API. That in turn lead to lots of issues that were
time consuming to triage and address.

This change makes Soong pass Metalava an empty previously released API
while disabling compatibility checks and API lint to replicate the
previous behavior.

This is preparation for making Metalava fail fast if it is provided
with API flags but no previously released API to which flagged APIs can
be reverted.

Bug: 425309457
Test: TH
Change-Id: Ibbe8a7fb1536bd066153653178294af082beba8b