Create a compatibility_test_suite_package module
This is to replace the all_compatibility_suites hardcoded list in soong.
Some test suites should only be created on certain manifests that have
all the dependencies they need, so if they're defined centrally in
soong they fail on minimal manifests. Creating a module that won't
be present on the minimal manifests solves this.

Bug: 425403103
Test: m catbox
Change-Id: Ie234da950276358b83f1f8876ed2ececdb63b633