Provide previously released API for unstable APIs, i.e. test
Previously, the `test` API did not provide a previously released API
as it is marked as `unstable`. The problem with that is it causes
problems when moving flagged APIs from `test` API to say `system` API
because there is no previous version of the API to revert to when the
flag is disabled.

This change adds a previously released API for `unstable` APIs but
maintains the existing behavior of not using it for API lint or
compatibility checking.

As `unstable` was the last API surface for which previously released
APIs were not provided this change means that previously released APIs
are always provided. That will allow Metalava to be strict and fail if
flagged APIs have no previously released API to which they can be
reverted.

Bug: 369373583
Test: ./gradlew
Change-Id: If983c1fe74b63053a685877df11b92ee83a8b403