Force recompilation when classes is missing.
Incremental Kotlin was failing when no changes to source were made,
but its `classes` output directory was missing.

This occured when incremental kotlin was turned off, reverting back
to non-incremental compilation, but leaving behind various incremental
artifacts. Non-incremental kotlin deletes the `classes` directory on
every invocation.

When incremental was turned back on, it saw the artifcacts, decided
that there was no work to do, and exited silently.

The follow up command depended upon the `classes` directory existing,
so it failed.

With this change, a missing output directory now causes a full
recompilation.

Fixes: 425280725
Test: delete `classes` and recompile
Change-Id: Idacbd5ac7e24d0805b8b51806616d3199281d5d5