Restrict the dependency tags in walking deps in addAutogeneratedRroDeps
Currently, there is not much restriction on dependency tag when
traversing the dependency in addAutogeneratedRroDeps. This may lead to
undesired rro dep from being installed in soong only builds. This change
restricts the type of dependency tags that are permitted to traverse the
dependency in finding the rro deps to install, to match the behavior in
make where the rro deps are installed via LOCAL_REQUIRED_MODULES
entries.

Test: build/soong/scripts/soong_only_diff_test.py cf_gwear_vanilla_x86
Bug: 425768106
Change-Id: Iee94a13c7e2fc25d0aa298aaf1f3b214a215920f