Define android_system_image_prebuilt
android_system_image_prebuilt module provides a prebuilt system image.
At the first stage it is implemented by inheriting soong 'prebuilt'
module, but we will change this to avoid the same module name with
the source-built images.

We still need to fill the correct values to fsInfo of the prebuilt
system image module because it is used for building the system_other
image. For now, system_other includes the dex files of non-system
partitions when using the prebuilt system image.

The android_system_image_prebuilt must include "file_contexts" for
the context of system_other files and "prop" to provide the prop file
used to build the prebuilt system image. However the prop file must
be removed and replaced with properties directly inject the
information to the android_system_image_prebuilt module.

Bug: 419170198
Test: Define a prebuilt system image and build
Change-Id: I1a3b9b5c3371e329e54bedb8e79280c55e1c349e