Add deps on tools referenced from compatability_test_suite_package
Currently, compatability_test_suite_package works based on installed
paths for tools. This is problematic for the licensing code, as it
needs more information from the module. We were finding the module
by searching the whole module graph for a module that produced the files
we're using, which is slow. Instead, add direct deps on those modules,
get the info we need from them, and use that to construct the notice
file.

Bug: 425403103
Test: Diff'd android-catbox.zip before/after, only difference was a little more information added to the notice file.
Change-Id: I5203d818a959ba463ef8935eaa64f0f5d1f07b6e